Move motion helpers to navigation
